+++ tests/common/Provider/ProviderTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Yiisoft\EventDispatcher\Tests\Provider;
 
-use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 use Yiisoft\EventDispatcher\Provider\ListenerCollection;
 use Yiisoft\EventDispatcher\Provider\Provider;
@@ -112,14 +111,6 @@ final class ProviderTest extends TestCase
 
         $this->assertContains('class interface', $interfaceHandlers);
         $this->assertContains('parent interface', $interfaceHandlers);
-    }
-
-    public function testListenerWithNoParameterThrowsException(): void
-    {
-        $this->expectException(InvalidArgumentException::class);
-        $this->expectExceptionMessage('Listeners must declare an object type they can accept.');
-
-        (new ListenerCollection())->add(fn () => null);
     }
 
     public function testListenerForEventIsReturned(): void